pyproject.toml
[tool.poetry]
name = "jama-cli"
version = "0.1.0"
description = "Command-line interface for Jama requirements management with MCP server support"
authors = ["Christian Nennemann <0lt7axgtr@mozmail.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/XORwell/jama-cli"
keywords = ["jama", "cli", "mcp", "model-context-protocol", "requirements-management", "api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
packages = [
    {include = "jama_cli"},
    {include = "jama_mcp_server"},  # Keep for backward compatibility
]
include = ["NOTICE"]

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.9.0"
py-jama-rest-client = "^1.17.0"
pydantic = "^2.0.0"
loguru = "^0.7.0"
typer = {version = ">=0.12.0", extras = ["all"]}
jsonschema = "^4.19.0"
python-dotenv = "^1.0.0"
mcp = "^1.0.0"
aiohttp-swagger3 = "^0.8.0"
pyyaml = "^6.0.0"
rich = "^13.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = "^24.0.0"
isort = "^5.13.0"
mypy = "^1.8.0"
ruff = "^0.1.0"
types-pyyaml = "^6.0.0"
pre-commit = "^3.6.0"
pytest-xdist = "^3.8.0"

[tool.poetry.scripts]
jama = "jama_cli.main:run"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["jama_cli", "jama_mcp_server"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_any_generics = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
strict_equality = true
strict_concatenate = true

[[tool.mypy.overrides]]
module = [
    "py_jama_rest_client.*",
    "aiohttp_swagger3.*",
    "mcp.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "jama_mcp_server.core.server",
    "jama_mcp_server.core.stdio_server",
]
disallow_untyped_defs = false
disable_error_code = ["union-attr"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
    "PLR",  # pylint refactoring
]
ignore = [
    "E501",    # line too long (handled by black)
    "B008",    # do not perform function calls in argument defaults
    "B905",    # zip without explicit strict
    "PLR0911", # too many return statements
    "PLR0912", # too many branches
    "PLR0913", # too many arguments (we have complex CLI commands)
    "PLR0915", # too many statements
    "PLR2004", # magic value comparison
]

[tool.ruff.per-file-ignores]
"tests/*" = ["ARG"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
# Default: fast unit tests only. Use --integration for full tests.
addopts = "-q --tb=short --ignore=tests/test_cli_integration.py --ignore=tests/test_read_operations.py --ignore=tests/test_write_operations.py --ignore=tests/test_mcp_server_integration.py"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::FutureWarning",
    "ignore:.*AppKey.*:UserWarning",
]
markers = [
    "integration: marks tests as integration tests (require live Jama instance)",
    "slow: marks tests as slow running",
]

[tool.coverage.run]
source = ["jama_cli", "jama_mcp_server"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
